comparing a float and double resulted in incorrect handling of dots (0 length elements) within dash patterns.
git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@1028 06663e23-700e-0410-b217-a244a6096597

+++ gs/src/gxpdash.c
@@ -56,7 +56,7 @@ subpath_expand_dashes(const subpath * psub, gx_path * ppath,
     const float *pattern = dash->pattern;
     int count, index;
     bool ink_on;
-    float elt_length;
+    double elt_length;
     fixed x0 = psub->pt.x, y0 = psub->pt.y;
     fixed x, y;
     const segment *pseg;
